==> Code/0_postag_model.ipynb <==
### Train a pos_tag model
- Using three datasets (Airbnb, Twitter, Yelp) <br>
- Keeps words and pos tags into a dictionary, where the key is a word and the value is a counter of POS and counts <br>
- For each word, only check if the most common pos_tag is equal. <br>

from nltk import pos_tag
import pandas as pd
import pickle, re
from sklearn.feature_extraction.text import CountVectorizer
from collections import defaultdict, Counter

airbnb_path = "/data/2/zwang/2018_S_WordTreatment/V2_airbnb/"
tw_path = "/data/2/zwang/2018_S_WordTreatment/V2_twitter/"
yp_path = "/data/2/zwang/2018_S_WordTreatment/V2_yelp/"

airbnb_corpus = "/data/2/zwang/2018_S_WordTreatment/V2_airbnb/1_Process/airbnb_labelby_sents_intotal.pickle"
tw_corpus = "/data/2/zwang/2018_S_WordTreatment/V2_twitter/1_Process/tw_labelby_sents_intotal.pickle"
yp_corpus = "/data/2/zwang/2018_S_WordTreatment/V2_yelp/1_Process/yp_labelby_sents_intotal.pickle"

def count_tags(airbnb_corpus,yp_corpus,tw_corpus,tag_file):
    word_tags = defaultdict(Counter)
    airbnb_pd = pd.DataFrame(pickle.load(open(airbnb_corpus,'rb')))
    yp_pd = pd.DataFrame(pickle.load(open(yp_corpus,'rb')))
    tw_pd = pd.DataFrame(pickle.load(open(tw_corpus,'rb')))
    
    for data_pd in [airbnb_pd,yp_pd,tw_pd]:
        print(data_pd.shape)
        for idx, row in data_pd.iterrows():
            if(idx % 10000 ==0):
                print(idx)
            tag_sent = pos_tag(row.sentence.split())
            for item in tag_sent:
                word_tags[item[0]][item[1]] += 1
    
    #pickle.dump(word_tags,open(tag_file,'wb'))
    return word_tags

#word_tags = count_tags(airbnb_corpus,yp_corpus,tw_corpus,tag_file='/data/2/zwang/2018_S_WordTreatment/V2_airbnb/word_tag_counter.pickle')

word_tags = pickle.load(open('/data/2/zwang/2018_S_WordTreatment/V2_airbnb/word_tag_counter.pickle','rb'))

word_tags['skinny']

word_tags['limited']

def store_word_tags(word_tags,treat_pairs_file,wd_tags_file):
    pair_pd = pd.read_csv(treat_pairs_file)
    vocab = list(set(pair_pd.source.values))+list(set(pair_pd.target.values))
    wd_tags = defaultdict(Counter)
    for wd in vocab:
        for tag_ct in word_tags[wd]:
            if(word_tags[wd][tag_ct]>20):
                wd_tags[wd][tag_ct] = word_tags[wd][tag_ct]
    
    pickle.dump(wd_tags,open(wd_tags_file,'wb'))
    return wd_tags

wd_tags = store_word_tags(word_tags,
                treat_pairs_file = '/data/2/zwang/2018_S_WordTreatment/V2_airbnb/1_Process/treat_pairs.csv',
                wd_tags_file = '/data/2/zwang/2018_S_WordTreatment/V2_airbnb/1_Process/wd_tags.pickle')

word_tags['canvas']

word_tags['display']

### Check if the most common pos_tag is equal

def check_com_pos(wd_tags_file,word_pairs_file,remain_wdpairs_file):
    wd_tags_pd = pickle.load(open(wd_tags_file,'rb'))
    wd_pairs_pd = pd.read_csv(word_pairs_file)
    remain_idx = []
    for idx,row in wd_pairs_pd.iterrows():
        if((row.source in wd_tags_pd) and (row.target in wd_tags_pd)):
            if(wd_tags_pd[row.source].most_common(1)[0][0] == wd_tags_pd[row.target].most_common(1)[0][0]):
                remain_idx.append(idx)
            else:
                print(row.source,wd_tags_pd[row.source].most_common(1),row.target,wd_tags_pd[row.target].most_common(1))
    
    wd_pairs_pd.iloc[remain_idx].to_csv(remain_wdpairs_file)
    print("%d remained, and %d removed" % (len(remain_idx),wd_pairs_pd.shape[0]-len(remain_idx)))

wd_tags_file = '/data/2/zwang/2018_S_WordTreatment/V2_airbnb/1_Process/wd_tags.pickle'
word_pairs_file = '/data/2/zwang/2018_S_WordTreatment/V2_airbnb/1_Process/treat_pairs_posinters.csv'
remain_wdpairs_file = '/data/2/zwang/2018_S_WordTreatment/V2_airbnb/1_Process/treat_pairs_posinters_poscom1.csv'
check_com_pos(wd_tags_file,word_pairs_file,remain_wdpairs_file)

wd_tags_file = '/data/2/zwang/2018_S_WordTreatment/V2_twitter/1_Process/wd_tags.pickle'
word_pairs_file = '/data/2/zwang/2018_S_WordTreatment/V2_twitter/1_Process/treat_pairs_posinters.csv'
remain_wdpairs_file = '/data/2/zwang/2018_S_WordTreatment/V2_twitter/1_Process/treat_pairs_posinters_poscom1.csv'
check_com_pos(wd_tags_file,word_pairs_file,remain_wdpairs_file)

wd_tags_file = '/data/2/zwang/2018_S_WordTreatment/V2_yelp/1_Process/wd_tags.pickle'
word_pairs_file = '/data/2/zwang/2018_S_WordTreatment/V2_yelp/1_Process/treat_pairs_posinters.csv'
remain_wdpairs_file = '/data/2/zwang/2018_S_WordTreatment/V2_yelp/1_Process/treat_pairs_posinters_poscom1.csv'
check_com_pos(wd_tags_file,word_pairs_file,remain_wdpairs_file)

### Train bigram model using all corpus data
- Keeps words and containing bigrams into a dictionary, where the key is a word and the value is a counter of bigrams and counts

def train_bigram(airbnb_corpus,yp_corpus,tw_corpus,bigram_file):
    word_bigram = {}
    airbnb_pd = pd.DataFrame(pickle.load(open(airbnb_corpus,'rb')))
    yp_pd = pd.DataFrame(pickle.load(open(yp_corpus,'rb')))
    tw_pd = pd.DataFrame(pickle.load(open(tw_corpus,'rb')))
    bi_vec = CountVectorizer(ngram_range=(2,2),stop_words=None,token_pattern='[a-z]+')
    
    for data_pd in [airbnb_pd,yp_pd,tw_pd]:
        print(data_pd.shape)
        X_data = bi_vec.fit_transform(data_pd.sentence.values)
        feat_ct = X_data.sum(axis=0).A1
        for feat in bi_vec.vocabulary_:
            if(feat not in word_bigram):
                word_bigram[feat] = 0
            word_bigram[feat] += feat_ct[bi_vec.vocabulary_[feat]]
        
    pickle.dump(word_bigram,open(bigram_file,'wb'))
    #return word_bigram

train_bigram(airbnb_corpus,yp_corpus,tw_corpus,bigram_file = '/data/2/zwang/2018_S_WordTreatment/V2_airbnb/bigram_dict.pickle')

word_bigram = pickle.load(open('/data/2/zwang/2018_S_WordTreatment/V2_airbnb/bigram_dict.pickle','rb'))

len(word_bigram)

list(word_bigram.keys())[:3]

word_bigram['see movies']

### Get bigram info for all word in the paraphrase vocabulary
- Keeps words and its bigrams into a dictionary, where the key is a word and the value is a counter of all containing bigrams and counts

def get_wd_bigram(all_bigram_file,wd_bigram_file):
    all_bigram = pickle.load(open(all_bigram_file,'rb'))
    wd_bigram = defaultdict(Counter)
    for bigram in all_bigram:
        wd1,wd2 = bigram.split()
        wd_bigram[wd1][bigram] += all_bigram[bigram]
        wd_bigram[wd2][bigram] += all_bigram[bigram]
    
    pickle.dump(wd_bigram,open(wd_bigram_file,'wb'))

all_bigram_file = '/data/2/zwang/2018_S_WordTreatment/V2_airbnb/bigram_dict.pickle'
wd_bigram_file = '/data/2/zwang/2018_S_WordTreatment/V2_airbnb/wd_bigram_dfdict.pickle'
get_wd_bigram(all_bigram_file,wd_bigram_file)

wd_bigram_dfdict = pickle.load(open(wd_bigram_file,'rb'))

type(wd_bigram_dfdict)

for item in wd_bigram_dfdict['otherwise']:
    print(item)
    break

def get_limitvocab_bigram(wd_bigram_file,airbnb_wdpair_file,tw_wdpair_file,yp_wdpair_file,vocab_bigram_file):
    wd_bigram_dfdict = pickle.load(open(wd_bigram_file,'rb'))
    vocab_ct = Counter()
    airbnb_pd = pd.read_csv(airbnb_wdpair_file)
    vocab_ct.update(list(airbnb_pd.source.values) + list(airbnb_pd.target.values))
    tw_pd = pd.read_csv(tw_wdpair_file)
    vocab_ct.update(list(tw_pd.source.values) + list(tw_pd.target.values))
    yp_pd = pd.read_csv(yp_wdpair_file)
    vocab_ct.update(list(yp_pd.source.values) + list(yp_pd.target.values))
    
    vocab_bigram = defaultdict(Counter)
    for wd in vocab_ct:
        for wd_bi in wd_bigram_dfdict[wd]:
            vocab_bigram[wd][wd_bi] += wd_bigram_dfdict[wd][wd_bi]
    
    pickle.dump(vocab_bigram,open(vocab_bigram_file,'wb'))
    print("Store bigram information for %d words." % len(vocab_ct))

wd_bigram_file = '/data/2/zwang/2018_S_WordTreatment/V2_airbnb/wd_bigram_dfdict.pickle'
airbnb_wdpair_file = '/data/2/zwang/2018_S_WordTreatment/V2_airbnb/1_Process/treat_pairs_posinters.csv'
tw_wdpair_file = '/data/2/zwang/2018_S_WordTreatment/V2_twitter/1_Process/treat_pairs_posinters.csv'
yp_wdpair_file = '/data/2/zwang/2018_S_WordTreatment/V2_yelp/1_Process/treat_pairs_posinters.csv'
vocab_bigram_file = '/data/2/zwang/2018_S_WordTreatment/V2_airbnb/vocab_bigram_dfdict.pickle'
get_limitvocab_bigram(wd_bigram_file,airbnb_wdpair_file,tw_wdpair_file,yp_wdpair_file,vocab_bigram_file)

- For a word pair, check all its containing sentences, and select those can be substituted with target word <br>
- check if after substitution by target word, the bigram exists in database <br>
- only maintain sentences that could be substituted by target word, and then divide into 10 folds to randomly select <br>

tw_path = '/data/2/zwang/2018_S_WordTreatment/V2_twitter/'
sort_file = tw_path+'3_CausalForest/tw_csf_200tree_treatment_posinters_sort.pickle'
csf_pd = pd.DataFrame(pickle.load(open(sort_file,'rb')))
csf_pd.shape

csf_pd.head()

csf_pd.ix[[1426624,1426675,1426681]]

def check_bigram_match(wdpair_sentence_file,word_pair_file,vocab_bigram_file,result_file):
    """
    Iterate over each word pair, and each sentence:
        check if the bigram exists after target word substitution:
        left_word control_word right_word --> left_word treatment_word right_word
        Only remain sentences that either (left_word treatment_word) or (treatment_word right_word) exist in treatment_word's history bigrams.
    return remaining word pairs, number of remaining sentences, and write sentences for each word pairs.
    """
    
    wdpair_pd = pd.DataFrame(pickle.load(open(wdpair_sentence_file,'rb')))
    wd_pairs = pd.read_csv(word_pair_file)
    vocab_bigram_dfdict = pickle.load(open(vocab_bigram_file,'rb'))
    remain_idx = []
    
    for idx,row in wdpair_pd.iterrows():
        # check if this word pair still exists in the limited vocabulary
        if(wd_pairs[(wd_pairs.source==row.source) & (wd_pairs.target==row.target)].shape[0]>0):
            word_list = re.findall('[a-z]+',row.sentence.lower())
            if(row.source in word_list):
                srci = word_list.index(row.source)
                tar_bigram_l = ''
                tar_bigram_r = ''
                if(srci > 0):
                    tar_bigram_l = word_list[srci-1]+' '+row.target
                if(srci < len(word_list)-1):
                    tar_bigram_r = row.target+' '+word_list[srci+1]

                flag = False
                if(tar_bigram_l in vocab_bigram_dfdict[row.target]):
                    flag = True
                elif(tar_bigram_r in vocab_bigram_dfdict[row.target]):
                    flag = True

                if(flag):
                    remain_idx.append(idx)

    print("%d remained out of %d in total" % (len(remain_idx),wdpair_pd.shape[0]))
    
    pickle.dump(wdpair_pd.ix[remain_idx],open(result_file,'wb'))

vocab_bigram_file = '/data/2/zwang/2018_S_WordTreatment/V2_airbnb/vocab_bigram_dfdict.pickle'
word_pair_file = tw_path + '1_Process/treat_pairs_posinters_poscom1.csv'

- results for same dataset should be the same

wdpair_sentence_file = tw_path+'3_CausalForest/tw_csf_200tree_treatment_posinters_sort.pickle'
result_file = tw_path+'3_CausalForest/tw_csf_200tree_treatment_posinters_sort_bigramcheck.pickle'
check_bigram_match(wdpair_sentence_file,word_pair_file,vocab_bigram_file,result_file)

wdpair_sentence_file = tw_path+'3_CounterFactual/tw_ctf_200tree_treatment_posinters_sort.pickle'
result_file = tw_path+'3_CounterFactual/tw_ctf_200tree_treatment_posinters_sort_bigramcheck.pickle'
check_bigram_match(wdpair_sentence_file,word_pair_file,vocab_bigram_file,result_file)

wdpair_sentence_file = tw_path+'3_VirtualTwins/tw_vt_200tree_treatment_posinters_sort.pickle'
result_file = tw_path+'3_VirtualTwins/tw_vt_200tree_treatment_posinters_sort_bigramcheck.pickle'
check_bigram_match(wdpair_sentence_file,word_pair_file,vocab_bigram_file,result_file)

wdpair_sentence_file = tw_path+'3_KNN/tw_knn_30_treatment_posinters_sort.pickle'
result_file = tw_path+'3_KNN/tw_csf_knn_30_treatment_posinters_sort_bigramcheck.pickle'
check_bigram_match(wdpair_sentence_file,word_pair_file,vocab_bigram_file,result_file)

def generate_wd_pairs(wdpair_bigram_file,original_wdpair_file,remain_wdpairs_file):
    wd_bigram_pd = pd.DataFrame(pickle.load(open(wdpair_bigram_file,'rb')))
    original_wdpairs = pd.read_csv(original_wdpair_file)
    remain_idx = []
    for idx,row in original_wdpairs.iterrows():
        if(wd_bigram_pd[(wd_bigram_pd.source == row.source) & (wd_bigram_pd.target == row.target)].shape[0]>0):
            remain_idx.append(idx)
        else:
            print(row.source,row.target)
            
    print("%d word pairs removed, %d remained" % (original_wdpairs.shape[0]-len(remain_idx),len(remain_idx)))
    original_wdpairs.iloc[remain_idx].to_csv(remain_wdpairs_file,index=False)

wdpair_bigram_file = airbnb_path+'3_KNN/airbnb_knn_30_treatment_posinters_sort_bigramcheck.pickle'
original_wdpair_file = airbnb_path +'1_Process/treat_pairs_posinters_poscom1.csv'
remain_wdpairs_file = airbnb_path +'1_Process/treat_pairs_posinters_poscom1_bigram.csv'
generate_wd_pairs(wdpair_bigram_file,original_wdpair_file,remain_wdpairs_file)

wdpair_bigram_file = tw_path+'3_KNN/tw_knn_30_treatment_posinters_sort_bigramcheck.pickle'
original_wdpair_file = tw_path +'1_Process/treat_pairs_posinters_poscom1.csv'
remain_wdpairs_file = tw_path +'1_Process/treat_pairs_posinters_poscom1_bigram.csv'
generate_wd_pairs(wdpair_bigram_file,original_wdpair_file,remain_wdpairs_file)

wdpair_bigram_file = yp_path+'3_KNN/yp_knn_30_treatment_posinters_sort_bigramcheck.pickle'
original_wdpair_file = yp_path +'1_Process/treat_pairs_posinters_poscom1.csv'
remain_wdpairs_file = yp_path +'1_Process/treat_pairs_posinters_poscom1_bigram.csv'
generate_wd_pairs(wdpair_bigram_file,original_wdpair_file,remain_wdpairs_file)

sent = "this weather is so Peaceful i luv it"
word_list = re.findall('[a-z]+',sent.lower())
print(word_list)
word_list.index('peaceful')

### If not, apply google n-gram
