==> Code/11_NewFeature.ipynb <==
### New low dimensional feature matrix
- Binary classifier vs 3-class classification <br>
> Binary: AMT rating increases(pos) vs others(neg)<br>
> 3-class: AMT rating increases(1), keep same(0), decreases(-1) <br>

- Coefficient of source word and target word <br>
- Number of support words and opposite words abs(coef) > 0.1 <br>
- Sentence length <br>
- Depth of source word and target word (VT,CTF,CSF) <br>
- Length of path for source sentence and target sentence <br>

import pickle, pydot, random, re, ast, time
import pandas as pd
import numpy as np
import ast
from scipy.sparse import vstack
from sklearn.externals.six import StringIO 
from IPython.display import display
#from sklearn.datasets import load_iris
from sklearn import tree
from IPython.display import Image
from collections import Counter
import seaborn as sns
from scipy.stats import pearsonr
from sklearn.ensemble import RandomForestClassifier
from sklearn.metrics.pairwise import cosine_similarity
from sklearn.feature_extraction.text import CountVectorizer, TfidfVectorizer
from sklearn.linear_model import LogisticRegression, LinearRegression
from sklearn.model_selection import cross_val_score
import matplotlib.pyplot as plt

%matplotlib inline

import warnings
warnings.filterwarnings("ignore")

project_path = '/data/2/zwang/2018_S_WordTreatment/'
amt_path = '/data/2/zwang/2018_S_WordTreatment/V2_AMT/6_amt/'
dataset_list = ['yp','tw','airbnb']
methods = ['KNN','VT','CTF','CSF']
full_name = {'yp':'Yelp','tw':'Twitter','airbnb':'Airbnb'}
data_file = {}
data_file['yp'] = project_path + 'V2_yelp/5_Select/yp_wdpair_sents_4effects_posinters_poscom1_bigramcheck_limit.pickle'
data_file['tw'] = project_path + 'V2_twitter/5_Select/tw_wdpair_sents_4effects_posinters_poscom1_bigramcheck_limit.pickle'
data_file['airbnb'] = project_path + 'V2_airbnb/5_Select/airbnb_wdpair_sents_4effects_posinters_poscom1_bigramcheck_limit_limitvocab.pickle'
vocab_file = project_path+'V2_airbnb/0_Data/common_wds.csv'

def cal_causal_words(data_pd,pos_thresh,neg_thresh):
    n_sup = []
    n_opp = []
    for sent in data_pd.sorted_sentence.values:
        nsp = 0
        nop = 0
        for item in ast.literal_eval(sent):
            if(item[1] > pos_thresh):
                nsp += 1
            elif(item[1] < neg_thresh):
                nop += 1
        n_sup.append(nsp)
        n_opp.append(nop)
    data_pd['n_sup'] = n_sup
    data_pd['n_opp'] = n_opp
    return data_pd

new_feats = ['src_coef','tar_coef','n_sup','n_opp','n_words','src_probLR','tar_probLR',
             'vt_srcwd_depth','vt_tarwd_depth','vt_src_pathlen','vt_tar_pathlen',
             'ctf_srcwd_depth','ctf_tarwd_depth','ctf_src_pathlen','ctf_tar_pathlen']

new_feats = ['tar_src_dif_coef','n_sup','n_opp','n_words','src_probLR']

def cal_newfeats(coef_pd,lr_file,vt_file,ctf_file):
    lr_pd = pd.read_csv(lr_file)
    lr_pd['n_words'] = lr_pd['src_sentence'].apply(lambda x: len(re.findall('\w+',x)))
    lr_pd['src_probLR'] = lr_pd['src_prob']
    lr_pd['tar_probLR'] = lr_pd['tar_prob']
    vt_pd = pd.read_csv(vt_file)
    vt_pd['vt_srcwd_depth'] = vt_pd['src_depth'].apply(lambda x: np.mean(ast.literal_eval(x)))
    vt_pd['vt_tarwd_depth'] = vt_pd['tar_depth'].apply(lambda x: np.mean(ast.literal_eval(x)))
    vt_pd['vt_src_pathlen'] = vt_pd['src_path_len'].apply(lambda x: np.mean(ast.literal_eval(x)))
    vt_pd['vt_tar_pathlen'] = vt_pd['tar_path_len'].apply(lambda x: np.mean(ast.literal_eval(x)))
    ctf_pd = pd.read_csv(ctf_file)
    ctf_pd['ctf_srcwd_depth'] = ctf_pd['src_depth'].apply(lambda x: np.mean(ast.literal_eval(x)))
    ctf_pd['ctf_tarwd_depth'] = ctf_pd['tar_depth'].apply(lambda x: np.mean(ast.literal_eval(x)))
    ctf_pd['ctf_src_pathlen'] = ctf_pd['src_path_len'].apply(lambda x: np.mean(ast.literal_eval(x)))
    ctf_pd['ctf_tar_pathlen'] = ctf_pd['tar_path_len'].apply(lambda x: np.mean(ast.literal_eval(x)))
            
    '''
    Has formatting problem when matching, make error on cases like 0.26666666666 and 0.26700000001.
    '''
    coef_lr_pd = coef_pd.merge(lr_pd,on=['id','source','target'])
    vt_ctf_pd = vt_pd.merge(ctf_pd,on=['id','source','target']) 
    super_pd = coef_lr_pd.merge(vt_ctf_pd,on=['id','source','target'])
    
    for col in ['amt_effect_mostcom','knn_effect','vt_effect','ctf_effect','csf_effect']:
        super_pd[col]=super_pd[col+'_x_x']
    return super_pd

def fit_LR_newfeats(dataset_list,new_feats,clf_flag):
    for di in range(len(dataset_list)):
        print("Processing %s ..." % (full_name[dataset_list[di]]))
        coef_file = project_path+'V2_'+full_name[dataset_list[di]].lower()+'/9_Coef/'+dataset_list[di]+'_sentence_coef.csv'
        org_coef_pd = pd.read_csv(coef_file)
        org_coef_pd['tar_src_dif_coef'] = org_coef_pd['tar_coef'] - org_coef_pd['src_coef'] 
        coef_pd = cal_causal_words(org_coef_pd,0.1,-0.1)
        lr_file = project_path+'V2_'+full_name[dataset_list[di]].lower()+'/8_Var/'+dataset_list[di]+'_effectVar_prob_byLR.csv'
        vt_file = project_path+'V2_'+full_name[dataset_list[di]].lower()+'/8_Var/'+dataset_list[di]+'_VT_srctar_depth3.csv'
        ctf_file = project_path+'V2_'+full_name[dataset_list[di]].lower()+'/8_Var/'+dataset_list[di]+'_CTF_srctar_depth.csv'
        super_pd = cal_newfeats(coef_pd,lr_file,vt_file,ctf_file)
        print(super_pd.shape)
        
        # construct feature matrix
        X_feats = []
        for i in range(super_pd.shape[0]):
            row = super_pd.iloc[i][new_feats].values
            X_feats.append([float(e) for e in row])
        
        # fit logistic-regression or linear-regression
        if(clf_flag == 'binary'):
            LR_clf = LogisticRegression(random_state=42,n_jobs=-1)
            amt_labels = super_pd['amt_effect_mostcom'].apply(lambda x: 1 if x>0 else 0)
            LR_clf.fit(X_feats,amt_labels)
            print(Counter(amt_labels),'%.3f' % (np.mean(cross_val_score(LR_clf, X_feats, amt_labels, cv=3))))
            for fi in LR_clf.coef_[0].argsort()[::-1]:
                print("%.3f , %s" % (LR_clf.coef_[0][fi],new_feats[fi]))
        
        elif(clf_flag == '3-class'):
            LR_clf = LogisticRegression(random_state=42,n_jobs=-1)
            amt_labels = super_pd['amt_effect_mostcom'].apply(lambda x: 1 if x>0 else (-1 if x<0 else 0))
            LR_clf.fit(X_feats,amt_labels)
            print(Counter(amt_labels),'%.3f' % (np.mean(cross_val_score(LR_clf, X_feats, amt_labels, cv=3))))
            for cls in LR_clf.classes_:
                print("class:",cls)
                ci = LR_clf.classes_[cls]
                for fi in LR_clf.coef_[ci].argsort()[::-1]:
                    print("%.3f , %s" % (LR_clf.coef_[ci][fi],new_feats[fi])) # 3-class
            
        elif(clf_flag == 'linear'):
            for md in ['knn_effect','vt_effect','ctf_effect','csf_effect']:
                print(md)
                LR_clf = LinearRegression(n_jobs=-1)
                LR_clf.fit(X_feats,super_pd[md].values)

                for idx in LR_clf.coef_.argsort()[::-1]:
                    print("%.3f , %s " % (LR_clf.coef_[idx], new_feats[idx]))

#### Result for 15 new features 

fit_LR_newfeats(dataset_list,new_feats,clf_flag='3-class')

fit_LR_newfeats(dataset_list,new_feats,clf_flag='binary')

fit_LR_newfeats(dataset_list,new_feats,clf_flag='linear')

#vt_pd['src_depth'].apply(lambda x: np.mean(ast.literal_eval(x)))

#### Result for 5 new features

fit_LR_newfeats(dataset_list,new_feats,clf_flag='3-class')

fit_LR_newfeats(dataset_list,new_feats,clf_flag='binary')

fit_LR_newfeats(dataset_list,new_feats,clf_flag='linear')

### Predict positive probability for each instance and then check correlation between P(y=1|x) and each treatment effect(KNN,VT,CTF,CSF).

def predict_prob_newfeats(dataset_list,new_feats,LSE_effect,flag):
    """
    Predict p(y=1|x) with new features.
    In-bag prediction
    """
    all_prob_corrs = []
    for di in range(len(dataset_list)):
        print("Processing %s ..." % (full_name[dataset_list[di]]))
        coef_file = project_path+'V2_'+full_name[dataset_list[di]].lower()+'/9_Coef/'+dataset_list[di]+'_sentence_coef.csv'
        org_coef_pd = pd.read_csv(coef_file)
        org_coef_pd['tar_src_dif_coef'] = org_coef_pd['tar_coef'] - org_coef_pd['src_coef'] 
        coef_pd = cal_causal_words(org_coef_pd,0.1,-0.1)
        lr_file = project_path+'V2_'+full_name[dataset_list[di]].lower()+'/8_Var/'+dataset_list[di]+'_effectVar_prob_byLR.csv'
        vt_file = project_path+'V2_'+full_name[dataset_list[di]].lower()+'/8_Var/'+dataset_list[di]+'_VT_srctar_depth3.csv'
        ctf_file = project_path+'V2_'+full_name[dataset_list[di]].lower()+'/8_Var/'+dataset_list[di]+'_CTF_srctar_depth.csv'
        super_pd = cal_newfeats(coef_pd,lr_file,vt_file,ctf_file)
        print(super_pd.shape)
        
        # construct feature matrix
        X_feats = []
        for i in range(super_pd.shape[0]):
            row = super_pd.iloc[i][new_feats].values
            X_feats.append([float(e) for e in row])
        X_feats = np.array(X_feats)
        
        # fit binary logistic-regression
        LR_clf = LogisticRegression(random_state=42,n_jobs=-1)
        amt_labels = super_pd[AMT_method].apply(lambda x: 1 if x>0 else 0)
        if(flag=='inb'):
            LR_clf.fit(X_feats,amt_labels)
            pos_prob = LR_clf.predict_proba(X_feats)[:,1]
        elif(flag=='oob'):
            pos_prob=[]
            for ti in range(X_feats.shape[0]):
                sel_idx = [idx for idx in range(X_feats.shape[0]) if idx != ti]
                LR_clf.fit(X_feats[sel_idx,:],amt_labels[sel_idx])
                pos_prob.append(LR_clf.predict_proba(X_feats[ti])[0,1])
                
        prob_corrs = []
        for eft in LSE_effect:
            prob_corrs.append(pearsonr(pos_prob,super_pd[eft])[0])
        all_prob_corrs.append(pd.DataFrame(prob_corrs,index=LSE_effect,columns=[full_name[dataset_list[di]]]))
        
        
    display(pd.concat(all_prob_corrs,axis=1))
        

LSE_effect = ['knn_effect','vt_effect','ctf_effect','csf_effect']
predict_prob_newfeats(dataset_list,new_feats,LSE_effect,AMT_method = 'amt_effect_mostcom',flag='inbag')

LSE_effect = ['knn_effect','vt_effect','ctf_effect','csf_effect']
predict_prob_newfeats(dataset_list,new_feats,LSE_effect,flag='oob')
