==> Code/12_rerunEstimators_AMTlabeled.ipynb <==
#### Re-run KNN, VT, CTF, CSF only for those appeared in AMT labeled data
- Re-run only for word pairs appear in AMT labeled file <br>
- KNN, VT, CTF, CSF\_200, CSF\_2000, CSF with new parameters, weighted\_KNN, combo method <br>
- Combine new treatment effects into AMT labeled file after re-run

import pickle,time,ast,csv
import pandas as pd
from nltk import pos_tag
import warnings
warnings.filterwarnings("ignore")

AMT_path = '/data/2/zwang/2018_S_WordTreatment/V2_AMT/'

def locate_AMT_sents(amt_pd,data_pd):
    """
    Locate the sentence id from original data file for each AMT labeled sentence.
    """
    sentid_list = []
    for idx,row in amt_pd.iterrows():
        wdpair_pd = data_pd[(data_pd['source'] == row['source']) & (data_pd['target'] == row['target'])]
        pair_sents = wdpair_pd['source_pos_sents'].values[0] + wdpair_pd['source_neg_sents'].values[0]
        try:
            sentid_list.append(pair_sents.index(row['src_sentence']))
        except:
#             return row,wdpair_pd
            sentid_list.append(-1)
    return sentid_list

def get_effect(effect_pd,amt_pd,method):
    effect_list = []
    for idx,row in amt_pd.iterrows():
        wdpair_pd = effect_pd[(effect_pd['source'] == row['source']) & (effect_pd['target'] == row['target'])]
        if(method != 'combo_effect'):
            pair_sents = ast.literal_eval(wdpair_pd['source_pos_sents_treatment'].values[0])[0] + ast.literal_eval(wdpair_pd['source_neg_sents_treatment'].values[0])[0]
        else:
            pair_sents = ast.literal_eval(wdpair_pd['source_pos_context_prob'].values[0])[0] + ast.literal_eval(wdpair_pd['source_neg_context_prob'].values[0])[0]
        effect_list.append(pair_sents[row['locate_id']][1])
    return effect_list

def get_wd_ratio(combo_pd,amt_pd):
    src_pos_ratio = []
    tar_pos_ratio = []
    for idx,row in amt_pd.iterrows():
        wdpair_pd = combo_pd[(combo_pd['source'] == row['source']) & (combo_pd['target'] == row['target'])]
        src_pos_ratio.append(wdpair_pd['src_pos_ratio'].values[0])
        tar_pos_ratio.append(wdpair_pd['tar_pos_ratio'].values[0])
    
    return src_pos_ratio,tar_pos_ratio

def combine_all_effects(AMT_file,data_file,result_file,effect_filelist):
    amt_pd = pd.read_csv(AMT_file,index_col=False)
    data_pd = pd.DataFrame(pickle.load(open(data_file,'rb')))
#     row,wdpair_pd = locate_AMT_sents(amt_pd,data_pd)
#     return row,wdpair_pd
    amt_pd['locate_id'] = locate_AMT_sents(amt_pd,data_pd)
    correct_pd = amt_pd[amt_pd['locate_id'] != -1]
    init_cols = list(amt_pd.columns)
#     return amt_pd
    
    for method in ['KNN','VT','CTF','CSF_200','CSF_2000','CSF_param','weighted_KNN','combo_effect']:
        effect_pd = pd.read_csv(effect_filelist[method],index_col=False)
#         return effect_pd
        correct_pd[method] = get_effect(effect_pd,correct_pd,method)
    
    combo_pd = pd.read_csv(effect_filelist['combo_effect'],index_col=False)
    src_pos_ratio,tar_pos_ratio = get_wd_ratio(combo_pd,correct_pd)
    correct_pd['src_pos_ratio'] = src_pos_ratio
    correct_pd['tar_pos_ratio'] = tar_pos_ratio

    correct_pd.to_csv(result_file,columns = init_cols+['KNN','VT','CTF','CSF_200','CSF_2000','CSF_param','weighted_KNN','combo_effect']+['src_pos_ratio','tar_pos_ratio'],
                  index = False)
    return correct_pd

prefix = 'tw'
effect_filemap = {}
effect_filemap['KNN'] = AMT_path+'AMT_WdSents/3_KNN/'+prefix+'_knn_30_treatment.csv'
effect_filemap['VT'] = AMT_path+'AMT_WdSents/3_VT/'+prefix+'_vt_200tree_treatment.csv'
effect_filemap['CTF'] = AMT_path+'AMT_WdSents/3_CTF/'+prefix+'_ctf_200tree_treatment.csv'
effect_filemap['CSF_200'] = AMT_path+'AMT_WdSents/3_CSF/'+prefix+'_csf_200tree_treatment.csv'
effect_filemap['CSF_2000'] = AMT_path+'AMT_WdSents/3_CSF/'+prefix+'_csf_2000tree_treatment.csv'
effect_filemap['CSF_param'] = AMT_path+'AMT_WdSents/3_CSF/'+prefix+'_csf_param_treatment.csv'
effect_filemap['weighted_KNN'] = AMT_path+'AMT_WdSents/3_KNN/'+prefix+'_weightedknn_30_treatment.csv'
effect_filemap['combo_effect'] = AMT_path+'AMT_WdSents/3_Combo/'+prefix+'_AMT_combo.csv'

start = time.time()
correct_pd = combine_all_effects(AMT_file = AMT_path+'6_amt/results/'+prefix+'_result_weightedAMT_nodummy.csv',
                             data_file = AMT_path+'AMT_WdSents/Data/'+prefix+'_AMT_wdsents.pickle',
                             result_file = AMT_path+'AMT_WdSents/Effects/'+prefix+'_AMT_comboEffects.csv',
                             effect_filelist = effect_filemap)
end = time.time()
print((end-start)/60)

correct_pd.head()

error_pd = amt_pd[amt_pd['locate_id'] == -1]
error_pd.shape

effect_pd = pd.read_csv(AMT_path+'AMT_WdSents/3_KNN/'+prefix+'_knn_30_treatment.csv',index_col=False)
effect_pd.shape

wdpair_pd = effect_pd[(effect_pd['source'] == 'famous') & (effect_pd['target'] == 'wonderful')]

ast.literal_eval(wdpair_pd['source_pos_sents_treatment'].values[0])[0]

effect_pd.head()

data_pd = pd.DataFrame(pickle.load(open(AMT_path+'AMT_WdSents/Data/'+prefix+'_AMT_wdsents.pickle','rb')))
wdpair_pd = data_pd[(data_pd['source'] == 'gay') & (data_pd['target'] == 'homo')]

print(error_pd.iloc[1]['src_sentence'])
print(pos_tag(error_pd.iloc[1]['src_sentence'].split()))

print(pos_tag('So obama was a homophobic teen who beat men he thot were homo . And if gay , then a rapist per Rachel #Zimmerman'.split()))

- Double check if mapped effects in each method is correct

prefix = 'airbnb'
effect_filemap = {}
effect_filemap['KNN'] = AMT_path+'AMT_WdSents/3_KNN/'+prefix+'_knn_30_treatment.csv'
effect_filemap['VT'] = AMT_path+'AMT_WdSents/3_VT/'+prefix+'_vt_200tree_treatment.csv'
effect_filemap['CTF'] = AMT_path+'AMT_WdSents/3_CTF/'+prefix+'_ctf_200tree_treatment.csv'
effect_filemap['CSF_200'] = AMT_path+'AMT_WdSents/3_CSF/'+prefix+'_csf_200tree_treatment.csv'
effect_filemap['CSF_2000'] = AMT_path+'AMT_WdSents/3_CSF/'+prefix+'_csf_2000tree_treatment.csv'
effect_filemap['CSF_param'] = AMT_path+'AMT_WdSents/3_CSF/'+prefix+'_csf_param_treatment.csv'
effect_filemap['weighted_KNN'] = AMT_path+'AMT_WdSents/3_KNN/'+prefix+'_weightedknn_30_treatment.csv'
effect_filemap['combo_effect'] = AMT_path+'AMT_WdSents/3_Combo/'+prefix+'_AMT_combo.csv'

src_wd = 'yummy'
tar_wd = 'delicious'
locate_id = 16
for md in effect_filemap:
    md_pd = pd.read_csv(effect_filemap[md],index_col=False)
    wdpair_pd = md_pd[(md_pd['source'] == src_wd) & (md_pd['target'] == tar_wd)]
    if(md != 'combo_effect'):
        all_sents = ast.literal_eval(wdpair_pd['source_pos_sents_treatment'].values[0])[0] + ast.literal_eval(wdpair_pd['source_neg_sents_treatment'].values[0])[0]
        print(md,all_sents[locate_id][1])
    elif(md == 'combo_effect'):
        all_sents = ast.literal_eval(wdpair_pd['source_pos_context_prob'].values[0])[0] + ast.literal_eval(wdpair_pd['source_neg_context_prob'].values[0])[0]
        print(md,all_sents[locate_id][1],wdpair_pd['src_pos_ratio'].values[0],wdpair_pd['tar_pos_ratio'].values[0])

wdpair_pd
